Create ghost.js
This code will scrape the title, heading, and paragraphs of a webpage and print them to the console. You can modify the code to select different elements or to do more advanced processing of the scraped data.

It's important to note that web scraping can be against the terms of service of some websites, and can also put a strain on their servers. Be sure to use web scraping responsibly and only scrape websites that allow it.